# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "hydra",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/recurring-rules/": {
      "post": {
        "tags": [
          "recurring-rules"
        ],
        "summary": "Create Recurring Rule",
        "description": "Create a recurring rule, such as rent or a subscription.\n\nArgs:\n    recurring_rule_service: The recurring rule service dependency.\n    household: The current household context.\n    rule_in: The rule to create.\n\nReturns:\n    The created rule, with the date it first falls due.\n\nRaises:\n    HTTPException: If an account or category does not exist in the household\n        (404), or the rule does not match its kind or would never come due\n        (400).",
        "operationId": "recurring-rules-create_recurring_rule",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecurringRuleCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecurringRulePublic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "recurring-rules"
        ],
        "summary": "List Recurring Rules",
        "description": "List the recurring rules of the household.\n\nArgs:\n    recurring_rule_service: The recurring rule service dependency.\n    household: The current household context.\n    is_active: An optional active state to filter on.\n    skip: Number of records to skip.\n    limit: Maximum number of records to return.\n\nReturns:\n    The rules, soonest due first.\n\nRaises:\n    HTTPException: If the user belongs to no household (404).",
        "operationId": "recurring-rules-list_recurring_rules",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "is_active",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Is Active"
            }
          },
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecurringRulesPublic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/recurring-rules/upcoming": {
      "get": {
        "tags": [
          "recurring-rules"
        ],
        "summary": "List Upcoming Occurrences",
        "description": "Show the recurring occurrences that have not been recorded yet.\n\nRead only: nothing is created, so this can be used to plan ahead.\n\nArgs:\n    recurring_rule_service: The recurring rule service dependency.\n    household: The current household context.\n    until: The last day to look ahead to, inclusive. Defaults to two months out.\n\nReturns:\n    The projected occurrences, soonest first, and what they add up to.\n\nRaises:\n    HTTPException: If the user belongs to no household (404).",
        "operationId": "recurring-rules-list_upcoming_occurrences",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "until",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Until"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpcomingOccurrencesPublic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/recurring-rules/run": {
      "post": {
        "tags": [
          "recurring-rules"
        ],
        "summary": "Run Recurring Rules",
        "description": "Record the transactions the rules have fallen due for.\n\nThe only way the rules are turned into transactions. The app calls it\nonce on load; it also serves for catching up by hand, or for creating\noccurrences ahead of today.\n\nArgs:\n    recurring_rule_service: The recurring rule service dependency.\n    household: The current household context.\n    until: The last day to record up to, inclusive. Defaults to today.\n\nReturns:\n    How many transactions were created and how many rules moved on.\n\nRaises:\n    HTTPException: If a rule points at an account that no longer exists (404).",
        "operationId": "recurring-rules-run_recurring_rules",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "until",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Until"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecurringRunResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/recurring-rules/{rule_id}": {
      "get": {
        "tags": [
          "recurring-rules"
        ],
        "summary": "Get Recurring Rule",
        "description": "Get one recurring rule of the household.\n\nArgs:\n    recurring_rule_service: The recurring rule service dependency.\n    household: The current household context.\n    rule_id: The ID of the rule.\n\nReturns:\n    The rule.\n\nRaises:\n    HTTPException: If the rule does not exist in the household (404).",
        "operationId": "recurring-rules-get_recurring_rule",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "rule_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Rule Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecurringRulePublic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "recurring-rules"
        ],
        "summary": "Update Recurring Rule",
        "description": "Edit a recurring rule, or pause it by setting is_active to false.\n\nArgs:\n    recurring_rule_service: The recurring rule service dependency.\n    household: The current household context.\n    rule_id: The ID of the rule to edit.\n    rule_in: The fields to change.\n\nReturns:\n    The updated rule.\n\nRaises:\n    HTTPException: If the rule or the category does not exist in the\n        household (404), or the new schedule would never come due (400).",
        "operationId": "recurring-rules-update_recurring_rule",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "rule_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Rule Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecurringRuleUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecurringRulePublic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "recurring-rules"
        ],
        "summary": "Delete Recurring Rule",
        "description": "Delete a recurring rule.\n\nThe transactions it already created are kept: they happened, and the rule\nis only the template that made them.\n\nArgs:\n    recurring_rule_service: The recurring rule service dependency.\n    household: The current household context.\n    rule_id: The ID of the rule to delete.\n\nReturns:\n    A confirmation message.\n\nRaises:\n    HTTPException: If the rule does not exist in the household (404).",
        "operationId": "recurring-rules-delete_recurring_rule",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "rule_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Rule Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Message": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "Message"
      },
...
      "RecurrenceFrequency": {
        "type": "string",
        "enum": [
          "weekly",
          "monthly",
          "yearly"
        ],
        "title": "RecurrenceFrequency"
      },
      "RecurringRuleCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "title": "Name"
          },
          "frequency": {
            "$ref": "#/components/schemas/RecurrenceFrequency",
            "default": "monthly"
          },
          "interval": {
            "type": "integer",
            "maximum": 1200.0,
            "minimum": 1.0,
            "title": "Interval",
            "default": 1
          },
          "start_date": {
            "type": "string",
            "format": "date",
            "title": "Start Date"
          },
          "end_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Date"
          },
          "day_of_month": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 31.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Day Of Month"
          },
          "kind": {
            "$ref": "#/components/schemas/TransactionKind",
            "default": "expense"
          },
          "amount_minor": {
            "type": "integer",
            "maximum": 4.611686018427388e+18,
            "exclusiveMinimum": 0.0,
            "title": "Amount Minor"
          },
          "merchant": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Merchant"
          },
          "note": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1024
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "default": true
          },
          "account_id": {
            "type": "string",
            "format": "uuid",
            "title": "Account Id"
          },
          "category_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category Id"
          },
          "counter_account_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Counter Account Id"
          }
        },
        "type": "object",
        "required": [
          "name",
          "start_date",
          "amount_minor",
          "account_id"
        ],
        "title": "RecurringRuleCreate"
      },
      "RecurringRulePublic": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "title": "Name"
          },
          "frequency": {
            "$ref": "#/components/schemas/RecurrenceFrequency",
            "default": "monthly"
          },
          "interval": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Interval",
            "default": 1
          },
          "start_date": {
            "type": "string",
            "format": "date",
            "title": "Start Date"
          },
          "end_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Date"
          },
          "day_of_month": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 31.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Day Of Month"
          },
          "kind": {
            "$ref": "#/components/schemas/TransactionKind",
            "default": "expense"
          },
          "amount_minor": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Amount Minor"
          },
          "merchant": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Merchant"
          },
          "note": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1024
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "default": true
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "household_id": {
            "type": "string",
            "format": "uuid",
            "title": "Household Id"
          },
          "account_id": {
            "type": "string",
            "format": "uuid",
            "title": "Account Id"
          },
          "category_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category Id"
          },
          "counter_account_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Counter Account Id"
          },
          "next_occurrence_on": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Occurrence On"
          },
          "last_generated_on": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Generated On"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "name",
          "start_date",
          "amount_minor",
          "id",
          "household_id",
          "account_id",
          "created_at"
        ],
        "title": "RecurringRulePublic"
      },
      "RecurringRuleUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "frequency": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RecurrenceFrequency"
              },
              {
                "type": "null"
              }
            ]
          },
          "interval": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 1200.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Interval"
          },
          "end_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Date"
          },
          "day_of_month": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 31.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Day Of Month"
          },
          "amount_minor": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 4.611686018427388e+18,
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Amount Minor"
          },
          "merchant": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Merchant"
          },
          "note": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1024
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "category_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category Id"
          },
          "is_active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Active"
          }
        },
        "type": "object",
        "title": "RecurringRuleUpdate"
      },
      "RecurringRulesPublic": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/RecurringRulePublic"
            },
            "type": "array",
            "title": "Data"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          }
        },
        "type": "object",
        "required": [
          "data",
          "count"
        ],
        "title": "RecurringRulesPublic"
      },
      "RecurringRunResult": {
        "properties": {
          "created_count": {
            "type": "integer",
            "title": "Created Count"
          },
          "skipped_count": {
            "type": "integer",
            "title": "Skipped Count"
          },
          "rules_advanced": {
            "type": "integer",
            "title": "Rules Advanced"
          }
        },
        "type": "object",
        "required": [
          "created_count",
          "skipped_count",
          "rules_advanced"
        ],
        "title": "RecurringRunResult",
        "description": "What one materialization pass did."
      },
...
      "TransactionKind": {
        "type": "string",
        "enum": [
          "expense",
          "income",
          "transfer"
        ],
        "title": "TransactionKind"
      },
...
      "UpcomingOccurrence": {
        "properties": {
          "rule_id": {
            "type": "string",
            "format": "uuid",
            "title": "Rule Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "kind": {
            "$ref": "#/components/schemas/TransactionKind"
          },
          "amount_minor": {
            "type": "integer",
            "title": "Amount Minor"
          },
          "occurs_on": {
            "type": "string",
            "format": "date",
            "title": "Occurs On"
          },
          "account_id": {
            "type": "string",
            "format": "uuid",
            "title": "Account Id"
          },
          "category_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category Id"
          }
        },
        "type": "object",
        "required": [
          "rule_id",
          "name",
          "kind",
          "amount_minor",
          "occurs_on",
          "account_id"
        ],
        "title": "UpcomingOccurrence",
        "description": "A date a rule will fall due on, not yet recorded."
      },
      "UpcomingOccurrencesPublic": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/UpcomingOccurrence"
            },
            "type": "array",
            "title": "Data"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          },
          "total_minor": {
            "type": "integer",
            "title": "Total Minor",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "data",
          "count"
        ],
        "title": "UpcomingOccurrencesPublic"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/login/access-token"
          }
        }
      }
    }
  }
}